+++ app/src/main/java/vn/asiantech/travelmate/extensions/Int.kt
@@ -4,8 +4,8 @@ import vn.asiantech.travelmate.utils.Constant
 import java.text.SimpleDateFormat
 import java.util.*
 
-fun Long.toDateFormat(): String = SimpleDateFormat(Constant.SIMPLE_DAY_FORMAT, Locale.US).format(
+fun Int.toDateFormat(): String = SimpleDateFormat(Constant.SIMPLE_DAY_FORMAT, Locale.US).format(
     Calendar.getInstance().apply {
-        timeInMillis = this@toDateFormat * 1000
+        timeInMillis = this@toDateFormat.toLong() * 1000
     }.time
 )
